scripts/requirements.txt
assemblyai>=0.23.0
openai>=1.54.0
httpx>=0.27.0
python-dotenv>=1.0.0
